refactor: moved `Head` tag, navbar and footer to `AppLayout` file
