+++ src/pages/_app.tsx
@@ -6,6 +6,7 @@ import { SWRConfig } from "swr";
 import theme from "@/styles/theme";
 
 import type { AppProps } from "next/app";
+import AppLayout from "@/components/AppLayout";
 export default function App({ Component, pageProps }: AppProps) {
     return (
         <ThemeProvider theme={theme}>
@@ -14,7 +15,9 @@ export default function App({ Component, pageProps }: AppProps) {
                     fetcher: (url: string) => fetch(url).then(res => res.json()),
                 }}
             >
-                <Component {...pageProps} />
+                <AppLayout>
+                    <Component {...pageProps} />
+                </AppLayout>
             </SWRConfig>
         </ThemeProvider>
     );
